Value attribute in PQV,  REAL should be of type Double instead of BigDecimal to be inconsistent with ISO xsd. Following classes are modified: gov.nih.nci.iso21090.Real gov.nih.nci.iso21090.Pqv gov.nih.nci.iso21090.grid.dto.transform.iso.PQTransformer gov.nih.nci.iso21090.grid.dto.transform.iso.REALTransformer gov.nih.nci.iso21090.grid.dto.transform.iso.IVLPQTransformerTest gov.nih.nci.iso21090.grid.dto.transform.iso.IVLREALTransformerTest gov.nih.nci.iso21090.grid.dto.transform.iso.PQTransformerTest gov.nih.nci.iso21090.grid.dto.transform.iso.REALTransformerTest gov.nih.nci.iso21090.RealTest
SVN-Revision: 195

+++ extensions/src/test/java/gov/nih/nci/iso21090/grid/dto/transform/iso/IVLPQTransformerTest.java
@@ -91,8 +91,6 @@ import gov.nih.nci.iso21090.Pq;
 import gov.nih.nci.iso21090.grid.dto.transform.AbstractTransformerTestBase;
 import gov.nih.nci.iso21090.grid.dto.transform.DtoTransformException;
 
-import java.math.BigDecimal;
-
 import org.iso._21090.IVLPQ;
 import org.iso._21090.PQ;
 import org.junit.Test;
@@ -152,7 +150,7 @@ public class IVLPQTransformerTest extends AbstractTransformerTestBase<IVLPQTrans
 
     private Pq getPq(double value) {
         Pq result = new PQTransformerTest().makeDtoSimple();
-        result.setValue(new BigDecimal(value));
+        result.setValue(new Double(value));
         return result;
     }
